Upgrade to new diagnostic provider API
